MAP-1448 LocationResource file split accordingly with Test split MAP-1402
