+++ src/main/kotlin/uk/gov/justice/digital/hmpps/locationsinsideprison/resource/LocationInt.kt
@@ -1,0 +1,3 @@
+package uk.gov.justice.digital.hmpps.locationsinsideprison.resource
+
+class LocationInt
