+++ src/main/kotlin/uk/gov/justice/digital/hmpps/locationsinsideprison/resource/LocationResource.kt
@@ -1,14 +1,11 @@
 package uk.gov.justice.digital.hmpps.locationsinsideprison.resource
 
 import io.swagger.v3.oas.annotations.Operation
-import io.swagger.v3.oas.annotations.Parameter
 import io.swagger.v3.oas.annotations.media.Content
 import io.swagger.v3.oas.annotations.media.Schema
 import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.validation.ValidationException
-import jakarta.validation.constraints.Max
-import jakarta.validation.constraints.PositiveOrZero
 import org.springdoc.core.annotations.ParameterObject
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
@@ -19,36 +16,19 @@ import org.springframework.http.MediaType
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PatchMapping
 import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.PutMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
-import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.Capacity
-import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.CreateNonResidentialLocationRequest
-import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.CreateResidentialLocationRequest
-import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.CreateWingRequest
 import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.LegacyLocation
-import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.LocationGroupDto
-import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.LocationPrefixDto
-import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.PatchNonResidentialLocationRequest
-import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.PatchResidentialLocationRequest
 import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.PermanentDeactivationLocationRequest
 import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.TemporaryDeactivationLocationRequest
 import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.UpdateLocationRequest
-import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.AccommodationType
-import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.ConvertedCellType
-import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.LocationType
-import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.NonResidentialUsageType
-import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.SpecialistCellType
-import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.UsedForType
 import uk.gov.justice.digital.hmpps.locationsinsideprison.service.InternalLocationDomainEventType
 import uk.gov.justice.digital.hmpps.locationsinsideprison.service.LocationService
-import uk.gov.justice.digital.hmpps.locationsinsideprison.service.ResidentialSummary
 import java.util.UUID
 import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.Location as LocationDTO
 
@@ -98,344 +78,8 @@ class LocationResource(
     @RequestParam(name = "includeChildren", required = false, defaultValue = "false") includeChildren: Boolean = false,
     @RequestParam(name = "includeHistory", required = false, defaultValue = "false") includeHistory: Boolean = false,
   ) =
-    locationService.getLocationById(id = id, includeChildren = includeChildren, includeHistory = includeHistory) ?: throw LocationNotFoundException(id.toString())
-
-  @GetMapping("/key/{key}")
-  @ResponseStatus(HttpStatus.OK)
-  @PreAuthorize("hasRole('ROLE_VIEW_LOCATIONS')")
-  @Operation(
-    summary = "Returns location information for this key",
-    description = "Requires role VIEW_LOCATIONS",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns location",
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the VIEW_LOCATIONS role",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Data not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun getLocationByKey(
-    @Schema(description = "Location Key", example = "MDI-A-1-001", required = true)
-    @PathVariable
-    key: String,
-    @RequestParam(name = "includeChildren", required = false, defaultValue = "false") includeChildren: Boolean = false,
-    @RequestParam(name = "includeHistory", required = false, defaultValue = "false") includeHistory: Boolean = false,
-  ) =
-    locationService.getLocationByKey(key = key, includeChildren = includeChildren, includeHistory = includeHistory) ?: throw LocationNotFoundException(key)
-
-  @PostMapping("/keys", produces = [MediaType.APPLICATION_JSON_VALUE])
-  @PreAuthorize("hasRole('ROLE_VIEW_LOCATIONS')")
-  @ResponseStatus(HttpStatus.OK)
-  @Operation(
-    summary = "Gets locations by their keys",
-    description = "Requires role VIEW_LOCATIONS",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns location",
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid Request",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the MAINTAIN_LOCATIONS role with write scope.",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Data not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun getLocationsByKeys(
-    @RequestBody
-    @Validated
-    keys: List<String>,
-  ): List<LocationDTO> = locationService.getLocationsByKeys(keys)
-
-  @GetMapping("/prison/{prisonId}")
-  @ResponseStatus(HttpStatus.OK)
-  @PreAuthorize("hasRole('ROLE_VIEW_LOCATIONS')")
-  @Operation(
-    summary = "Return locations for this prison",
-    description = "Requires role VIEW_LOCATIONS",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns locations",
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the VIEW_LOCATIONS role",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Data not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun getLocationForPrison(
-    @Schema(description = "Prison Id", example = "MDI", required = true, minLength = 3, maxLength = 5, pattern = "^[A-Z]{2}I|ZZGHI$")
-    @PathVariable
-    prisonId: String,
-  ): List<LocationDTO> = locationService.getLocationByPrison(prisonId)
-
-  @GetMapping("/prison/{prisonId}/groups")
-  @ResponseStatus(HttpStatus.OK)
-  @PreAuthorize("hasRole('ROLE_VIEW_LOCATIONS')")
-  @Operation(
-    summary = "Return location groups for this prison",
-    description = "Requires role VIEW_LOCATIONS",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns location groups",
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the VIEW_LOCATIONS role",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Data not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun getLocationGroupsForPrison(
-    @Schema(description = "Prison Id", example = "MDI", required = true, minLength = 3, maxLength = 5, pattern = "^[A-Z]{2}I|ZZGHI$")
-    @PathVariable
-    prisonId: String,
-  ): List<LocationGroupDto> = locationService.getLocationGroupsForPrison(prisonId)
-
-  @GetMapping("/prison/{prisonId}/group/{group}/location-prefix")
-  @PreAuthorize("hasRole('ROLE_VIEW_LOCATIONS')")
-  @Operation(
-    summary = "Get location prefix by group.",
-    description = "Requires role VIEW_LOCATIONS",
-    responses = [
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid request.",
-        content =
-        [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the VIEW_LOCATIONS role",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-
-      ApiResponse(
-        responseCode = "404",
-        description = "Data not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "500",
-        description = "Unrecoverable error occurred whilst processing request.",
-        content =
-        [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun getLocationPrefixFromGroup(
-    @Schema(description = "Prison Id", example = "MDI", required = true, minLength = 3, maxLength = 5, pattern = "^[A-Z]{2}I|ZZGHI$")
-    @PathVariable("prisonId")
-    prisonId: String,
-    @Schema(description = "The group name", example = "Houseblock 1", required = true)
-    @PathVariable("group")
-    group: String,
-  ): LocationPrefixDto =
-    locationService.getLocationPrefixFromGroup(prisonId, group)
-
-  @GetMapping("/groups/{prisonId}/{name}")
-  @Operation(
-    summary = "List of cell locations by group at prison location.",
-    description = "Requires role VIEW_LOCATIONS",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns cell locations by group",
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the VIEW_LOCATIONS role",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Data not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  @ResponseStatus(HttpStatus.OK)
-  @PreAuthorize("hasRole('ROLE_VIEW_LOCATIONS')")
-  fun getLocationGroup(
-    @Schema(description = "Prison Id", example = "MDI", required = true, minLength = 3, maxLength = 5, pattern = "^[A-Z]{2}I|ZZGHI$")
-    @PathVariable
-    prisonId: String,
-    @Parameter(description = "Group name", required = true)
-    @PathVariable("name")
-    name: String,
-  ) =
-    locationService.getCellLocationsForGroup(prisonId, name)
-
-  @GetMapping("/prison/{prisonId}/archived")
-  @ResponseStatus(HttpStatus.OK)
-  @PreAuthorize("hasRole('ROLE_VIEW_LOCATIONS')")
-  @Operation(
-    summary = "Return residential archived locations for this prison",
-    description = "Requires role VIEW_LOCATIONS",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns archived locations",
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the VIEW_LOCATIONS role",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Data not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun getArchivedLocationForPrison(
-    @Schema(description = "Prison Id", example = "MDI", required = true, minLength = 3, maxLength = 5, pattern = "^[A-Z]{2}I|ZZGHI$")
-    @PathVariable
-    prisonId: String,
-  ) = locationService.getArchivedLocations(prisonId)
-
-  @GetMapping("/prison/{prisonId}/inactive-cells")
-  @ResponseStatus(HttpStatus.OK)
-  @PreAuthorize("hasRole('ROLE_VIEW_LOCATIONS')")
-  @Operation(
-    summary = "Return residential inactive cells for this prison",
-    description = "Requires role VIEW_LOCATIONS",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns inactive locations",
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the VIEW_LOCATIONS role",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun getInactiveLocationsForPrison(
-    @Schema(description = "Prison Id", example = "MDI", required = true, minLength = 3, maxLength = 5, pattern = "^[A-Z]{2}I|ZZGHI$")
-    @PathVariable
-    prisonId: String,
-    @Schema(description = "location ID below which all inactive cells will be returned", example = "de91dfa7-821f-4552-a427-bf2f32eafeb0", required = false)
-    @RequestParam(name = "parentLocationId", required = false)
-    parentLocationId: UUID? = null,
-  ) = locationService.getResidentialInactiveLocations(prisonId, parentLocationId)
-
-  @GetMapping("/residential-summary/{prisonId}")
-  @ResponseStatus(HttpStatus.OK)
-  @PreAuthorize("hasRole('ROLE_VIEW_LOCATIONS')")
-  @Operation(
-    summary = "Return locations for this prison below the parent location, is not provided - top level (w.g. WINGS)",
-    description = "Requires role VIEW_LOCATIONS",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns locations for this level",
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the VIEW_LOCATIONS role",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Data not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun getLocationForPrisonBelowParent(
-    @Schema(description = "Prison Id", example = "MDI", required = true, minLength = 3, maxLength = 5, pattern = "^[A-Z]{2}I|ZZGHI$")
-    @PathVariable
-    prisonId: String,
-    @Schema(description = "Parent location ID", example = "de91dfa7-821f-4552-a427-bf2f32eafeb0", required = false)
-    @RequestParam(name = "parentLocationId", required = false)
-    parentLocationId: UUID? = null,
-    @Schema(description = "Parent location path hierarchy, can be a Wing code, or landing code", example = "A-1", required = false)
-    @RequestParam(name = "parentPathHierarchy", required = false)
-    parentPathHierarchy: String? = null,
-    @Schema(description = "Include latest history", required = false, defaultValue = "false")
-    @RequestParam(name = "latestHistory", required = false, defaultValue = "false")
-    latestHistory: Boolean = false,
-  ): ResidentialSummary = locationService.getResidentialLocations(
-    prisonId = prisonId,
-    parentLocationId = parentLocationId,
-    parentPathHierarchy = parentPathHierarchy,
-    returnLatestHistory = latestHistory,
-  )
+    locationService.getLocationById(id = id, includeChildren = includeChildren, includeHistory = includeHistory)
+      ?: throw LocationNotFoundException(id.toString())
 
   @GetMapping("")
   @PreAuthorize("hasRole('ROLE_VIEW_LOCATIONS')")
@@ -475,355 +119,6 @@ class LocationResource(
       throw ValidationException("Page size must be 200 or less")
     }
     return locationService.getLocations(pageable)
-  }
-
-  @GetMapping("/prison/{prisonId}/non-residential-usage-type/{usageType}")
-  @ResponseStatus(HttpStatus.OK)
-  @PreAuthorize("hasRole('ROLE_VIEW_LOCATIONS')")
-  @Operation(
-    summary = "Return non-residential locations by usage for this prison",
-    description = "Requires role VIEW_LOCATIONS",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns non-residential locations",
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid Request",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the VIEW_LOCATIONS role",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun getLocationsByPrisonAndNonResidentialUsageType(
-    @Schema(description = "Prison Id", example = "MDI", required = true, minLength = 3, maxLength = 5, pattern = "^[A-Z]{2}I|ZZGHI$")
-    @PathVariable
-    prisonId: String,
-    @Schema(description = "Usage type", example = "APPOINTMENTS", required = true)
-    @PathVariable
-    usageType: NonResidentialUsageType,
-  ): List<LocationDTO> = locationService.getLocationsByPrisonAndNonResidentialUsageType(prisonId, usageType)
-
-  @PostMapping("/residential", produces = [MediaType.APPLICATION_JSON_VALUE])
-  @PreAuthorize("hasRole('ROLE_MAINTAIN_LOCATIONS') and hasAuthority('SCOPE_write')")
-  @ResponseStatus(HttpStatus.CREATED)
-  @Operation(
-    summary = "Creates a residential location",
-    description = "Requires role MAINTAIN_LOCATIONS and write scope",
-    responses = [
-      ApiResponse(
-        responseCode = "201",
-        description = "Returns created location",
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid Request",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the MAINTAIN_LOCATIONS role with write scope.",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Data not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "409",
-        description = "Location already exists",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun createResidentialLocation(
-    @RequestBody
-    @Validated
-    createResidentialLocationRequest: CreateResidentialLocationRequest,
-  ): LocationDTO {
-    return eventPublishAndAudit(
-      InternalLocationDomainEventType.LOCATION_CREATED,
-    ) {
-      locationService.createResidentialLocation(createResidentialLocationRequest)
-    }
-  }
-
-  @PostMapping("/create-wing", produces = [MediaType.APPLICATION_JSON_VALUE])
-  @PreAuthorize("hasRole('ROLE_MAINTAIN_LOCATIONS') and hasAuthority('SCOPE_write')")
-  @ResponseStatus(HttpStatus.CREATED)
-  @Operation(
-    summary = "Creates a residential wing with landings and cells",
-    description = "Requires role MAINTAIN_LOCATIONS and write scope",
-    responses = [
-      ApiResponse(
-        responseCode = "201",
-        description = "Returns created locations",
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid Request",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the MAINTAIN_LOCATIONS role with write scope.",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "409",
-        description = "Location already exists",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun createWing(
-    @RequestBody
-    @Validated
-    createWingRequest: CreateWingRequest,
-  ): LocationDTO {
-    return eventPublishAndAudit(
-      InternalLocationDomainEventType.LOCATION_CREATED,
-    ) {
-      locationService.createWing(createWingRequest)
-    }
-  }
-
-  @PostMapping("/non-residential", produces = [MediaType.APPLICATION_JSON_VALUE])
-  @PreAuthorize("hasRole('ROLE_MAINTAIN_LOCATIONS') and hasAuthority('SCOPE_write')")
-  @ResponseStatus(HttpStatus.CREATED)
-  @Operation(
-    summary = "Creates a non-residential location",
-    description = "Requires role MAINTAIN_LOCATIONS and write scope",
-    responses = [
-      ApiResponse(
-        responseCode = "201",
-        description = "Returns created location",
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid Request",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the MAINTAIN_LOCATIONS role with write scope.",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Data not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "409",
-        description = "Location already exists",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun createNonResidentialLocation(
-    @RequestBody
-    @Validated
-    createNonResidentialLocationRequest: CreateNonResidentialLocationRequest,
-  ): LocationDTO {
-    return eventPublishAndAudit(
-      InternalLocationDomainEventType.LOCATION_CREATED,
-    ) {
-      locationService.createNonResidentialLocation(createNonResidentialLocationRequest)
-    }
-  }
-
-  @PatchMapping("/residential/{id}")
-  @PreAuthorize("hasRole('ROLE_MAINTAIN_LOCATIONS') and hasAuthority('SCOPE_write')")
-  @Operation(
-    summary = "Partial update of a location",
-    description = "Requires role MAINTAIN_LOCATIONS and write scope",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns updated location",
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid Request",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the MAINTAIN_LOCATIONS role with write scope.",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Data not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "409",
-        description = "Location already exists",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun patchResidentialLocation(
-    @Schema(description = "The location Id", example = "de91dfa7-821f-4552-a427-bf2f32eafeb0", required = true)
-    @PathVariable
-    id: UUID,
-    @RequestBody
-    @Validated
-    patchLocationRequest: PatchResidentialLocationRequest,
-  ): LocationDTO {
-    val results = locationService.updateResidentialLocation(id, patchLocationRequest)
-    eventPublish(buildEvents(results))
-    audit(id.toString()) { results.location.copy(childLocations = null, parentLocation = null) }
-    return results.location
-  }
-
-  @PatchMapping("/non-residential/{id}")
-  @PreAuthorize("hasRole('ROLE_MAINTAIN_LOCATIONS') and hasAuthority('SCOPE_write')")
-  @Operation(
-    summary = "Partial update of a non-residential location",
-    description = "Requires role MAINTAIN_LOCATIONS and write scope",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns updated location",
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid Request",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the MAINTAIN_LOCATIONS role with write scope.",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Data not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "409",
-        description = "Location already exists",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun patchNonResidentialLocation(
-    @Schema(description = "The location Id", example = "de91dfa7-821f-4552-a427-bf2f32eafeb0", required = true)
-    @PathVariable
-    id: UUID,
-    @RequestBody
-    @Validated
-    patchLocationRequest: PatchNonResidentialLocationRequest,
-  ): LocationDTO {
-    val results = locationService.updateNonResidentialLocation(id, patchLocationRequest)
-    eventPublish(buildEvents(results))
-    audit(id.toString()) { results.location.copy(childLocations = null, parentLocation = null) }
-    return results.location
-  }
-
-  fun buildEvents(results: UpdateLocationResult): () -> Map<InternalLocationDomainEventType, List<LocationDTO>> {
-    val locationsChanged = if (results.otherParentLocationChanged != null) {
-      listOf(results.location, results.otherParentLocationChanged)
-    } else {
-      listOf(results.location)
-    }
-
-    return {
-      mapOf(
-        InternalLocationDomainEventType.LOCATION_AMENDED to locationsChanged,
-      )
-    }
-  }
-
-  @PutMapping("/{id}/capacity")
-  @PreAuthorize("hasRole('ROLE_MAINTAIN_LOCATIONS') and hasAuthority('SCOPE_write')")
-  @Operation(
-    summary = "Set the capacity of a cell",
-    description = "Requires role MAINTAIN_LOCATIONS and write scope",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns cell location changes",
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid Request",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the MAINTAIN_LOCATIONS role with write scope.",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Location not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun changeCapacity(
-    @Schema(description = "The location Id", example = "de91dfa7-821f-4552-a427-bf2f32eafeb0", required = true)
-    @PathVariable
-    id: UUID,
-    @RequestBody
-    @Validated
-    capacity: Capacity,
-  ): LocationDTO {
-    return eventPublishAndAudit(
-      InternalLocationDomainEventType.LOCATION_AMENDED,
-    ) {
-      locationService.updateCellCapacity(
-        id = id,
-        maxCapacity = capacity.maxCapacity,
-        workingCapacity = capacity.workingCapacity,
-      )
-    }
   }
 
   @PutMapping("/{id}")
@@ -1066,7 +361,11 @@ class LocationResource(
     @Schema(description = "The location Id", example = "de91dfa7-821f-4552-a427-bf2f32eafeb0", required = true)
     @PathVariable
     id: UUID,
-    @RequestParam(name = "cascade-reactivation", required = false, defaultValue = "false") reactivateSubLocations: Boolean = false,
+    @RequestParam(
+      name = "cascade-reactivation",
+      required = false,
+      defaultValue = "false",
+    ) reactivateSubLocations: Boolean = false,
   ): LocationDTO {
     return eventPublishAndAudit(
       InternalLocationDomainEventType.LOCATION_REACTIVATED,
@@ -1074,277 +373,7 @@ class LocationResource(
       locationService.reactivateLocation(id, reactivateSubLocations)
     }
   }
-
-  @PutMapping("/{id}/convert-cell-to-non-res-cell")
-  @PreAuthorize("hasRole('ROLE_MAINTAIN_LOCATIONS') and hasAuthority('SCOPE_write')")
-  @Operation(
-    summary = "Converts a cell to a non res cell location",
-    description = "Requires role MAINTAIN_LOCATIONS and write scope",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns updated location",
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid Request",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the MAINTAIN_LOCATIONS role with write scope.",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Location not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun convertCellToNonResidentialLocation(
-    @Schema(description = "The location Id", example = "de91dfa7-821f-4552-a427-bf2f32eafeb0", required = true)
-    @PathVariable
-    id: UUID,
-    @RequestBody
-    @Validated
-    convertCellToNonResidentialLocationRequest: ConvertCellToNonResidentialLocationRequest,
-  ): LocationDTO {
-    return eventPublishAndAudit(
-      InternalLocationDomainEventType.LOCATION_AMENDED,
-    ) {
-      with(convertCellToNonResidentialLocationRequest) {
-        locationService.convertToNonResidentialCell(id, convertedCellType, otherConvertedCellType)
-      }
-    }
-  }
-
-  @PutMapping("/{id}/convert-to-cell")
-  @PreAuthorize("hasRole('ROLE_MAINTAIN_LOCATIONS') and hasAuthority('SCOPE_write')")
-  @Operation(
-    summary = "Makes a non-res cell location a cell",
-    description = "Requires role MAINTAIN_LOCATIONS and write scope",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns updated location",
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid Request",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the MAINTAIN_LOCATIONS role with write scope.",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Location not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "409",
-        description = "Location Accommodation Type Other Non Residential cannot be accepted",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun convertToCellLocation(
-    @Schema(description = "The location Id", example = "de91dfa7-821f-4552-a427-bf2f32eafeb0", required = true)
-    @PathVariable
-    id: UUID,
-    @RequestBody
-    @Validated
-    convertToCellRequest: ConvertToCellRequest,
-  ): LocationDTO {
-    return eventPublishAndAudit(
-      InternalLocationDomainEventType.LOCATION_AMENDED,
-    ) {
-      if (convertToCellRequest.accommodationType == AccommodationType.OTHER_NON_RESIDENTIAL) {
-        throw LocationCannotBeResidentialException(convertToCellRequest.accommodationType.toString())
-      }
-      with(convertToCellRequest) {
-        locationService.convertToCell(
-          id = id,
-          accommodationType = accommodationType,
-          specialistCellType = specialistCellType,
-          maxCapacity = maxCapacity,
-          workingCapacity = workingCapacity,
-          usedForTypes = usedForTypes,
-        )
-      }
-    }
-  }
-
-  @PutMapping("/{id}/used-for-type", produces = [MediaType.APPLICATION_JSON_VALUE])
-  @PreAuthorize("hasRole('ROLE_MAINTAIN_LOCATIONS') and hasAuthority('SCOPE_write')")
-  @Operation(
-    summary = "Update Used For types at cell location below the specified location",
-    description = "Requires role MAINTAIN_LOCATIONS and write scope",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns details of the updated locations",
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid Request",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the MAINTAIN_LOCATIONS role with write scope.",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Location not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun updateUsedForType(
-    @Schema(description = "The location Id", example = "de91dfa7-821f-4552-a427-bf2f32eafeb0", required = true)
-    @PathVariable
-    id: UUID,
-    @RequestBody
-    @Validated
-    usedFor: Set<UsedForType>,
-  ): LocationDTO {
-    return eventPublishAndAudit(
-      InternalLocationDomainEventType.LOCATION_AMENDED,
-    ) {
-      locationService.updateResidentialLocationUsedForTypes(id = id, usedFor = usedFor)
-    }
-  }
-
-  @PutMapping("/{id}/specialist-cell-types", produces = [MediaType.APPLICATION_JSON_VALUE])
-  @PreAuthorize("hasRole('ROLE_MAINTAIN_LOCATIONS') and hasAuthority('SCOPE_write')")
-  @Operation(
-    summary = "Update specialist cell types for a cell",
-    description = "Requires role MAINTAIN_LOCATIONS and write scope",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns details of the updated cell",
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid Request",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the MAINTAIN_LOCATIONS role with write scope.",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Location not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun updateSpecialistCellTypes(
-    @Schema(description = "The location Id", example = "de91dfa7-821f-4552-a427-bf2f32eafeb0", required = true)
-    @PathVariable
-    id: UUID,
-    @RequestBody
-    @Validated
-    specialistCellTypes: Set<SpecialistCellType>,
-  ): LocationDTO {
-    return eventPublishAndAudit(
-      InternalLocationDomainEventType.LOCATION_AMENDED,
-    ) {
-      locationService.updateSpecialistCellTypes(id = id, specialistCellTypes = specialistCellTypes)
-    }
-  }
-
-  @GetMapping("/prison/{prisonId}/location-type/{locationType}")
-  @ResponseStatus(HttpStatus.OK)
-  @PreAuthorize("hasRole('ROLE_VIEW_LOCATIONS')")
-  @Operation(
-    summary = "Return locations by their type for this prison",
-    description = "Requires role VIEW_LOCATIONS",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Returns locations",
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Missing required role. Requires the VIEW_LOCATIONS role",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Data not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun getLocationsByPrisonAndLocationType(
-    @Schema(description = "Prison Id", example = "MDI", required = true, minLength = 3, maxLength = 5, pattern = "^[A-Z]{2}I|ZZGHI$")
-    @PathVariable
-    prisonId: String,
-    @Schema(description = "Location type", example = "CELL", required = true)
-    @PathVariable
-    locationType: LocationType,
-  ): List<LocationDTO> = locationService.getLocationByPrisonAndLocationType(prisonId, locationType)
 }
-
-@Schema(description = "Request to convert a cell to a non-res location")
-data class ConvertCellToNonResidentialLocationRequest(
-  @Schema(description = "Cell type to convert to", example = "SHOWER", required = true)
-  val convertedCellType: ConvertedCellType,
-  @Schema(description = "Other type of converted cell", example = "Swimming pool", required = false)
-  val otherConvertedCellType: String? = null,
-)
-
-@Schema(description = "Request to convert a non-res location to a cell")
-data class ConvertToCellRequest(
-  @Schema(description = "Accommodation type of the location", example = "NORMAL_ACCOMMODATION", required = true)
-  val accommodationType: AccommodationType,
-  @Schema(description = "Specialist cell type", example = "BIOHAZARD_DIRTY_PROTEST", required = false)
-  val specialistCellType: SpecialistCellType?,
-  @Schema(description = "Max capacity", example = "2", required = true)
-  @field:Max(value = 99, message = "Max capacity cannot be greater than 99")
-  @field:PositiveOrZero(message = "Max capacity cannot be less than 0")
-  val maxCapacity: Int = 0,
-  @Schema(description = "Working capacity", example = "1", required = true)
-  @field:Max(value = 99, message = "Working capacity cannot be greater than 99")
-  @field:PositiveOrZero(message = "Working capacity cannot be less than 0")
-  val workingCapacity: Int = 0,
-  @Schema(description = "Used For list", example = "STANDARD_ACCOMMODATION, PERSONALITY_DISORDER", required = false)
-  val usedForTypes: List<UsedForType>? = null,
-)
 
 data class UpdateLocationResult(
   val location: LocationDTO,
